# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nourish API
  description: >
    AI-powered nutrition & food tracking API. Every route except /health and
    /api/auth/register|login requires a Bearer JWT and is scoped to the
    authenticated user. Import this file directly into Postman, Insomnia, or
    Swagger UI (swagger.io/tools/swagger-ui) to explore and call it.
  version: "1.0.0"
servers:
  - url: http://localhost:4000/api
    description: Local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: invalid_request }
            message: { type: string }
            details: { type: object, nullable: true }
...
  responses:
...
    NotFound:
      description: Not found (or not owned by the caller — existence is never leaked)
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Invalid:
      description: Failed validation
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
paths:
...
  /profile:
    get:
      summary: Get profile + goals
      responses:
        "200": { description: OK }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      summary: Update profile fields (partial)
      requestBody:
        content: { application/json: { schema: { type: object } } }
      responses:
        "200": { description: OK }
        "400": { $ref: "#/components/responses/Invalid" }
  /profile/goals:
    put:
      summary: Set daily nutrition targets directly
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [calories, proteinG, carbsG, fatG, fibreG]
              properties:
                calories: { type: integer }
                proteinG: { type: number }
                carbsG: { type: number }
                fatG: { type: number }
                fibreG: { type: number }
      responses:
        "200": { description: OK }
  /profile/goals/recalculate:
    post:
      summary: Recompute targets from profile (Mifflin-St Jeor) — always overridable, not medical advice
      responses:
        "200": { description: OK }
        "400": { description: Profile missing age/height/weight }
...
  /recipes:
    get:
      summary: List your saved recipes with computed totals
      responses: { "200": { description: OK } }
    post:
      summary: Create a recipe from ingredients (each qty/unit resolved via the food catalog)
      requestBody:
        required: true
        content: { application/json: { schema: { type: object } } }
      responses:
        "201": { description: Created }
  /recipes/{id}:
    delete:
      summary: Delete a recipe
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /recipes/{id}/log:
    post:
      summary: Log N servings of a saved recipe as a new meal
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [date, mealType]
              properties:
                servingsConsumed: { type: number, default: 1 }
                date: { type: string, format: date }
                mealType: { type: string }
                time: { type: string }
      responses:
        "201": { description: Created }
  /weight:
    get:
      summary: List weight logs (optional start/end range)
      parameters:
        - { name: start, in: query, schema: { type: string, format: date } }
        - { name: end, in: query, schema: { type: string, format: date } }
      responses: { "200": { description: OK } }
    post:
      summary: Log (or update) today's — or any date's — weight
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [date, weightKg]
              properties:
                date: { type: string, format: date }
                weightKg: { type: number }
                note: { type: string }
      responses: { "201": { description: Created/updated } }
  /water:
    get:
      summary: Get water logs + total for a date
      parameters: [{ name: date, in: query, required: true, schema: { type: string, format: date } }]
      responses: { "200": { description: OK } }
    post:
      summary: Log a water entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [date, amountMl]
              properties:
                date: { type: string, format: date }
                amountMl: { type: integer }
      responses: { "201": { description: Created } }
...
  /notifications/{id}/read:
    post:
      summary: Mark one notification read
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "204": { description: OK } }
  /notifications/{id}:
    delete:
      summary: Delete a notification
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "204": { description: Deleted } }
...
  /notifications/preferences:
    get:
      summary: Get notification preferences
      responses: { "200": { description: OK } }
    put:
      summary: Set one notification preference (type, enabled, and/or time)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type: { type: string, enum: [mealReminders, hydration, loggingReminder, goalProgress, dailyReport, weeklyReport] }
                enabled: { type: boolean }
                time: { type: string, example: "20:00" }
      responses: { "200": { description: OK } }
  /notifications/devices:
    post:
      summary: Register a device push token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pushToken, platform]
              properties: { pushToken: { type: string }, platform: { type: string, enum: [ios, android, web] } }
      responses: { "201": { description: Created } }
  /reports/daily:
    get:
      summary: Get the daily report as JSON
      parameters: [{ name: date, in: query, required: true, schema: { type: string, format: date } }]
      responses: { "200": { description: OK } }
  /reports/daily/download:
    get:
      summary: Download the daily report as an HTML file
      parameters: [{ name: date, in: query, required: true, schema: { type: string, format: date } }]
      responses:
        "200":
          description: OK
          content: { text/html: { schema: { type: string } } }
  /reports/weekly:
    get:
      summary: Get the weekly report as JSON (defaults to the 7 days ending today)
      parameters: [{ name: end, in: query, schema: { type: string, format: date } }]
      responses: { "200": { description: OK } }
  /reports/weekly/download:
    get:
      summary: Download the weekly report as an HTML file
      parameters: [{ name: end, in: query, schema: { type: string, format: date } }]
      responses: { "200": { description: OK } }
  /reports/daily/send:
    post:
      summary: Send today's (or a given date's) report by email right now
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [date], properties: { date: { type: string, format: date } } }
      responses: { "200": { description: OK } }
...